% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year.R
\name{add_mid_year}
\alias{add_mid_year}
\title{Add Mid-Year Column}
\usage{
add_mid_year(data, year_var = dx_year, sep = "-")
}
\description{
Adds a new column containing the midpoint year of the range given in the
column indicated by \code{year_var}.
}
\seealso{
Other year processors: \code{\link{complete_year_groups}}
}
\concept{year processors}
