% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age_adjust}
\alias{age_adjust}
\title{Age-Adjust Rates}
\usage{
age_adjust(data, count = n, population = fcds::seer_pop_fl,
  population_standard = fcds::seer_std_ages, by_year = c(dx_year_mid =
  "year"), age = age_group, keep_age = FALSE)
}
\arguments{
\item{data}{A data frame, containing counts}

\item{count}{The unquoted column name containing raw event or outcome counts.}

\item{population}{Population data specific to the population described by
\code{data}. By default, uses the county-specific Florida population data
provided by SEER (see \link{seer_pop_fl}).}

\item{population_standard}{The standard age-specific population used to
calculate the age-adjusted rates. By default, uses the 2000 U.S. standard
population provided by SEER (see \link{seer_std_ages}).}

\item{keep_age}{Age-adjustment by definition summarizes event or outcome
counts (incidence) over \emph{all} included ages. Set \code{keep_age} to join the
source data with all age-specific population without completing the age
adjustment calculation. This option is primarily provided for debugging
purposes.}

\item{year}{The unquoted column name containing the year that will be matched
to the population data.}
}
\value{
A data frame with age-adjusted incidence rates. Note that \code{age_group}
will no longer be included in the output because the age-adjusted rate
summarizes the observed incidence across all ages.
}
\description{
Provides age-adjusted incident rates with respect to a reference population,
by default \link{seer_std_ages}.
}
\section{Age-Adjusted Rates}{


Calculating age-adjusted rates requires three primary inputs:
\enumerate{
\item Raw age-specific count of event or outcome, possibly observed or
summarized at repeated, consistent time intervals.
\item Population data with the same demographic resolution as the age-specific
counts, as in, for example, the population for the same geographic region,
sex, race, year, and age.
\item The standard reference population that is used to weight incidence among
the observed age-specific count.
}

Note that each input is required to have a column \code{age_group} containing
matching age groups across all three inputs.
}

