% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcds.R
\name{age_boundaries}
\alias{age_boundaries}
\title{Convert Age Group Column to Low and High Age Range}
\usage{
age_boundaries(.data, age_var = age_group, sep = "\\\\s*-\\\\s*")
}
\arguments{
\item{.data}{A data frame}

\item{age_var}{Unquoted column name containing the age grouping.}

\item{sep}{Separator between columns.

If character, is interpreted as a regular expression. The default
value is a regular expression that matches any sequence of
non-alphanumeric values.

If numeric, interpreted as positions to split at. Positive values start
at 1 at the far-left of the string; negative value start at -1 at the
far-right of the string. The length of \code{sep} should be one less than
\code{into}.}
}
\description{
Takes the age group stored in \code{age_var} and creates two variables,
\code{age_low} and \code{age_high} with the age boundaries of the group.
}
\seealso{
Other age processors: \code{\link{complete_age_groups}},
  \code{\link{filter_age}}
}
\concept{age processors}
