% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcds.R
\name{complete_age_groups}
\alias{complete_age_groups}
\title{Complete Age Groups}
\usage{
complete_age_groups(.data, low = NULL, high = NULL, fill = list(n =
  0), include_unknown = FALSE)
}
\arguments{
\item{.data}{A data frame}

\item{low}{Low age boundary (inclusive)}

\item{high}{High age boundary (inclusive)}

\item{fill}{A named list that for each variable supplies a single value to
use instead of \code{NA} for missing combinations.}

\item{include_unknown}{Should the "Unknown" age group be included?}
}
\description{
Completes age groups by adding missing age groups, either within the age
range from \code{low} to \code{high} or using the full age list from the SEER data.
}
\seealso{
Other age processors: \code{\link{expand_age_groups}},
  \code{\link{filter_age}}
}
\concept{age processors}
