% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year.R
\name{complete_year_groups}
\alias{complete_year_groups}
\title{Complete Year Groups}
\usage{
complete_year_groups(.data, start = NULL, end = NULL,
  year_var = "dx_year", fill = list(n = 0))
}
\description{
Completes year groups with expected years.
}
\seealso{
Other year processors: \code{\link{add_mid_year}}
}
\concept{year processors}
