% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcds-data.R
\docType{data}
\name{county_fips_fl}
\alias{county_fips_fl}
\title{Florida County FIPS Codes}
\format{A data frame with 68 rows and 2 variables: \describe{
\item{\code{county_name}}{Name of the county in Florida.}
\item{\code{fips_code}}{FIPS code.} }}
\usage{
county_fips_fl
}
\description{
Federal Information Processing Standards (FIPS) County Codes for Florida, as
defined in \code{20 Appendix B FIPS County Codes for Florida.pdf} provided by
FCDS.
}
\keyword{datasets}
