% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fcds_cache}
\alias{fcds_cache}
\alias{fcds_cache_ls}
\alias{fcds_cache_info}
\alias{fcds_cache_clean}
\title{List or Clean Cached FCDS Data Files}
\usage{
fcds_cache_ls(path = NULL, pattern = "^fcds_.+\\\\.[Rr][Dd][Ss]$")

fcds_cache_info(path = NULL, ...)

fcds_cache_clean(path = NULL, ..., all = FALSE, dry_run = FALSE)
}
\arguments{
\item{path}{The path to the fcds cached data, defaults to
\code{\link[=fcds_default_data_path]{fcds_default_data_path()}}}

\item{pattern}{The file pattern used to identify fcds cached data files,
by default these are expected to be \code{.rds} files whose names start with
\code{fcds_}.}

\item{...}{Arguments passed on to \code{fcds_cache_ls}
\describe{
  \item{pattern}{The file pattern used to identify fcds cached data files,
by default these are expected to be \code{.rds} files whose names start with
\code{fcds_}.}
}}

\item{all}{Should all cached data files be removed? If \code{FALSE}, the default
value, the most recent file is kept. Recency is determined by sorting file
names in reverse order rather than by file system modification times. If in
doubt, use \code{dry_run = TRUE} to check behavior.}

\item{dry_run}{If \code{TRUE}, \code{fcds_cache_clean()} reports planned actions
without removing any files.}
}
\description{
Helper functions to list cached processed \strong{fcds} data files or to clean
outdated cached files.
}
\section{Functions}{
\itemize{
\item \code{fcds_cache_ls}: List cached fcds data files

\item \code{fcds_cache_info}: Report information about cached fcds data files

\item \code{fcds_cache_clean}: Clean outdated cached fcds data files
}}

\seealso{
Other FCDS Import Functions: \code{\link{fcds_default_data_path}},
  \code{\link{fcds_import}}, \code{\link{fcds_load}},
  \code{\link{fcds_recoding}}
}
\concept{FCDS Import Functions}
