% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fcds_const}
\alias{fcds_const}
\title{List Expected or Valid FCDS Constants}
\usage{
fcds_const(var = "year_group", full = FALSE,
  fcds_recoding_file = NULL)
}
\arguments{
\item{var}{An FCDS variable or a package constant. Set to \code{NULL} for a list
of valid variables.}

\item{full}{If \code{FALSE}, returns only the values that are expected in the
cleaned FCDS data. If \code{TRUE}, returns information regarding the original
variable name (\code{name_original}) and value (\code{value}) and the cleaned
variable name (\code{name_clean}) and value label (\code{label}).}

\item{fcds_recoding_file}{The path to the recoding specification yaml file.
Set to \code{NULL} to use the default recoding as used in \code{\link[=fcds_import]{fcds_import()}}.}
}
\value{
A character vector of valid FCDS values, or a tibble with information
about the original and recoded FCDS data values.
}
\description{
Lists the expected values in the processed, cached FCDS data using the
built-in recoding, or alternatively returns a tibble containing information
about the cleaned column and value labels and the original data values.
}
\examples{

fcds_const("year_group")

fcds_const("year_group", full = TRUE)

fcds_const("cancer_site_group", full = TRUE) \%>\% head()

fcds_const("moffitt_catchment")

}
