% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fcds_default_data_path}
\alias{fcds_default_data_path}
\title{Default Path for FCDS Data}
\usage{
fcds_default_data_path(path = NULL)
}
\arguments{
\item{path}{Set the default path for storing or looking for cached FCDS data.}
}
\value{
Returns the current global setting for the path to the FCDS data
cache.
}
\description{
Returns the default data path for FCDS data. You can globally set the path
using the \code{fcds.default_data_path} option, or by providing a path to
\code{fcds_default_data_path()}. Restore the FCDS package default cache by setting
the default to \code{NULL}.
}
\examples{
fcds_default_data_path()

# Set to ~/tmp
fcds_default_data_path("~/tmp")
fcds_default_data_path()

# Restore global default
fcds_default_data_path(NULL)
fcds_default_data_path()

}
\seealso{
Other FCDS Import Functions: \code{\link{fcds_import}},
  \code{\link{fcds_load}}
}
\concept{FCDS Import Functions}
