% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcds-data.R
\docType{data}
\name{fcds_example}
\alias{fcds_example}
\title{FCDS Example Data}
\format{A data frame with 7,951 rows and 36 variables. Note that the official
data contains more than 3 million rows. This example, while representative
of the offical data, contains simulated data for only 15 counties in
Florida, may not include all cancer types or all possible values seen in
the data and has absolutely no overlap (i.e. no row-wise matches, excluding
\code{patient_id}) with the published data.}
\usage{
fcds_example
}
\description{
Example data similar to the processed FCDS data that is returned by
\code{\link[=fcds_import]{fcds_import()}}. The values contained within are only representative of the
types of values that appear in the final data set, but have \strong{no overlap}
whatsoever with the published data.
}
\keyword{datasets}
