% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fcds_import}
\alias{fcds_import}
\title{Import FCDS Data}
\usage{
fcds_import(file, output_dir = fcds_default_data_path(),
  output_file = "fcds_\%F-\%H\%M.rds", ...,
  keep_original_columns = FALSE, fcds_recoding = NULL,
  verbose = TRUE, col_types = readr::cols(.default =
  readr::col_character()))
}
\arguments{
\item{file}{The raw FCDS data file}

\item{output_dir}{The location where the cleaned FCDS data files should be
located. Set to \code{NULL} save to current working directory. By default,
stores the processed data in \code{\link[=fcds_default_data_path]{fcds_default_data_path()}}, which in general
points to the global FCDS data cache.}

\item{output_file}{Name of the file to store the cleaned FCDS data file. If
you don't want to save the cleaned data, set \code{output_file = NULL}. The file
name is formatted using \code{\link[=strftime]{strftime()}} and the current system time.}

\item{...}{Arguments passed on to \code{readr::read_csv}
\describe{
  \item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector).

Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
be automatically uncompressed. Files starting with \code{http://},
\code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
downloaded. Remote gz files can also be automatically downloaded and
decompressed.

Literal data is most useful for examples and tests. It must contain at
least one new line to be recognised as data (instead of a path) or be a
vector of greater than length 1.

Using a value of \code{\link[=clipboard]{clipboard()}} will read from the system clipboard.}
  \item{quote}{Single character used to quote strings.}
  \item{col_names}{Either \code{TRUE}, \code{FALSE} or a character vector
of column names.

If \code{TRUE}, the first row of the input will be used as the column
names, and will not be included in the data frame. If \code{FALSE}, column
names will be generated automatically: X1, X2, X3 etc.

If \code{col_names} is a character vector, the values will be used as the
names of the columns, and the first row of the input will be read into
the first row of the output data frame.

Missing (\code{NA}) column names will generate a warning, and be filled
in with dummy names \code{X1}, \code{X2} etc. Duplicate column names
will generate a warning and be made unique with a numeric prefix.}
  \item{col_types}{One of \code{NULL}, a \code{\link[=cols]{cols()}} specification, or
a string. See \code{vignette("readr")} for more details.

If \code{NULL}, all column types will be imputed from the first 1000 rows
on the input. This is convenient (and fast), but not robust. If the
imputation fails, you'll need to supply the correct types yourself.

If a column specification created by \code{\link[=cols]{cols()}}, it must contain
one column specification for each column. If you only want to read a
subset of the columns, use \code{\link[=cols_only]{cols_only()}}.

Alternatively, you can use a compact string representation where each
character represents one column:
c = character, i = integer, n = number, d = double,
l = logical, f = factor, D = date, T = date time, t = time, ? = guess, or
\code{_}/\code{-} to skip the column.}
  \item{locale}{The locale controls defaults that vary from place to place.
The default locale is US-centric (like R), but you can use
\code{\link[=locale]{locale()}} to create your own locale that controls things like
the default time zone, encoding, decimal mark, big mark, and day/month
names.}
  \item{na}{Character vector of strings to interpret as missing values. Set this
option to \code{character()} to indicate no missing values.}
  \item{quoted_na}{Should missing values inside quotes be treated as missing
values (the default) or strings.}
  \item{comment}{A string used to identify comments. Any text after the
comment characters will be silently ignored.}
  \item{trim_ws}{Should leading and trailing whitespace be trimmed from
each field before parsing it?}
  \item{skip}{Number of lines to skip before reading data.}
  \item{n_max}{Maximum number of records to read.}
  \item{guess_max}{Maximum number of records to use for guessing column types.}
  \item{progress}{Display a progress bar? By default it will only display
in an interactive session and not while knitting a document. The display
is updated every 50,000 values and will only display if estimated reading
time is 5 seconds or more. The automatic progress bar can be disabled by
setting option \code{readr.show_progress} to \code{FALSE}.}
  \item{skip_empty_rows}{Should blank rows be ignored altogether? i.e. If this
option is \code{TRUE} then blank rows will not be represented at all.  If it is
\code{FALSE} then they will be represented by \code{NA} values in all the columns.}
}}

\item{keep_original_columns}{Should the original FCDS columns be kept in the
imported data? By default, only the cleaned columns are kept.}

\item{fcds_recoding}{The FCDS recoding definition. See the FCDS Recoding
section for more information. Set to \code{NULL} to use the fcds package
default settings.}

\item{verbose}{Prints additional information about the importing process}

\item{col_types}{Passed to \code{\link[readr:read_csv]{readr::read_csv()}}. By default, all columns are
imported initially as character and are recoded or updated during the
pre-processing step.}
}
\value{
A tibble containing the pre-processed FCDS and optionally the
original columns of the raw FCDS data (if \code{keep_original_columns = TRUE}).
The pre-processing step provides the following columns.
\itemize{
\item \code{patient_id}: Patient ID Number.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#20}{20}.
Derived from \code{Patient_ID_N20}.
\item \code{year_group}: Year of Diagnosis (5 year group).
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#390}{390}.
Derived from \code{Date_of_Dx_Year_Recoded}.
\item \code{year}: Year of Diagnosis (midpoint of 5 year group).
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#390}{390}.
Derived from \code{Date_of_Dx_Year_Recoded}.
\item \code{cancer_status}: Cancer Status at time abstract was completed.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#1770}{1770}.
Derived from \code{Cancer_Status_N1770}.
\item \code{cancer_site_group}: FCDS Site Group.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#2220}{2220}.
Derived from \code{FCDS_Site_Group}.
\item \code{cancer_site_specific}: FCDS Site Group with specific within-group areas.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#2220}{2220}.
Derived from \code{FCDS_Site_Group}.
\item \code{confirmation}: Diagnostic Confirmation at first diagnosis.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#490}{490}.
Derived from \code{Diagnostic_Confirmation_N490}.
\item \code{age_group}: FCDS Age Group.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#2220}{2220}.
Derived from \code{FCDS_Age_Group}.
\item \code{race}: Race (recoded).
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#160}{160}.
Derived from \code{Race_Recoded}.
\item \code{sex}: Sex (recoded).
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#220}{220}.
Derived from \code{Sex_Recoded}.
\item \code{origin}: Spanish/Hispanic Origin (recoded).
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#190}{190}.
Derived from \code{Ethnicity_Recoded}.
\item \code{marital_status}: Marital Status at diagnosis (recoded).
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#150}{150}.
Derived from \code{Marital_Status_Recoded}.
\item \code{county_name}: County Name of patient's primary residence at the time tumor was diagnosed.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#90}{90}.
Derived from \code{County_at_DX_N90}.
\item \code{county_fips}: County FIPS Code of patient's primary residence at the time tumor was diagnosed.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#90}{90}.
Derived from \code{County_at_DX_N90}.
\item \code{state}: State of patient's primary residence at the time of diagnosis (recoded).
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#80}{80}.
Derived from \code{Addr_at_DX_State_Recoded}.
\item \code{florida_resident}: Patient's primary state of residence was Florida at time of diagnosis.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#80}{80}.
Derived from \code{Addr_at_DX_State_Recoded}.
\item \code{country}: Country of patient's primary residence at time of diagnosis (recoded).
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#102}{102}.
Derived from \code{Addr_at_Dx_Country_Recoded}.
\item \code{birth_country}: Country of Birthplace (recoded).
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#254}{254}.
Derived from \code{Birthplace_Country_Recoded}.
\item \code{birth_state}: State of Birthplace (recoded).
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#254}{254}.
Derived from \code{Birthplace_State_Abrv_Recoded}.
\item \code{primary_payer}: Primary Payer at Diagnosis (recoded).
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#630}{630}.
Derived from \code{Dx_Primary_Payor_Recoded}.
\item \code{cancer_reporting_source}: Type of Reporting Source.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#500}{500}.
Derived from \code{Type_of_Reporting_Source_N500}.
\item \code{cancer_ICDO3_conversion}: ICD-O-3 Conversion Flag.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#2116}{2116}.
Derived from \code{ICDO3_Conversion_FL_N2116}.
\item \code{cancer_laterality}: Laterality at Diagnosis.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#410}{410}.
Derived from \code{Laterality_N410}.
\item \code{cancer_grade}: Grade, Differentiation, or Cell Lineage Indicator (SEER/CCCR).
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#440}{440}.
Derived from \code{Grade_N440}.
\item \code{cancer_ICDO3_histology}: Histologic Type ICD-O-3.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#522}{522}.
Derived from \code{Histologic_Type_ICDO3_N522}.
\item \code{cancer_ICDO3_behavior}: Behavior Code ICD-O-3.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#523}{523}.
Derived from \code{Behavior_Code_ICDO3_N523}.
\item \code{cancer_ICDO3_morphology}: Morphology Code ICD-O-3 (Type and Behavior).
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#521}{521}.
Derived from \code{Histologic_Type_ICDO3_N522}, \code{Behavior_Code_ICDO3_N523}.
\item \code{seer_stage_1977}: SEER Summary Stage 1977.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#760}{760}.
Derived from \code{SEER_Summ_Stage_1977_N760}.
\item \code{seer_stage_2000}: SEER Summary Stage 2000.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#759}{759}.
Derived from \code{SEER_Summ_Stage_2000_N759}.
\item \code{seer_stage}: SEER Stage from 2000 falling back to 1977.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#759}{759}.
Derived from \code{seer_stage_1977}, \code{seer_stage_2000}.
\item \code{seer_stage_derived_1977}: Derivation of SEER Summary Stage 1977.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#3040}{3040}.
Derived from \code{Derived_SS1977_FL_N3040}.
\item \code{seer_stage_derived_2000}: Derivation of SEER Summary Stage 2000.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#3050}{3050}.
Derived from \code{Derived_SS2000_FL_N3050}.
\item \code{tobacco_cigarette}: Cigarette smoking.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#9965}{9965}.
Derived from \code{FCDS_Tob_Use_Cigarette_N1300}.
\item \code{tobacco_other}: Smoking tobacco products other than cigarettes (e.g., pipes, cigars, kreteks).
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#9966}{9966}.
Derived from \code{FCDS_Tob_Use_OthSmoke_N1300}.
\item \code{tobacco_smokeless}: Smokeless tobacco products (e.g, chewing tobacco, snuff, etc.).
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#9967}{9967}.
Derived from \code{FCDS_Tob_Use_Smokeless_Tob_N1300}.
\item \code{tobacco_nos}: Tobacco NOS, includes use of e-cigarettes and vaporizers.
NAACCR Item #\href{http://datadictionary.naaccr.org/default.aspx?c=10#9968}{9968}.
Derived from \code{FCDS_Tob_Use_NOS_N1300}.
}
}
\description{
Imports the FCDS data file and applies basic data pre-processing. If the user
provides only the path to the raw FCDS data, by default this function will
save a cached version of the pre-processed data in \link{fcds} default directory
(see \code{\link[=fcds_default_data_path]{fcds_default_data_path()}} for more information). The user can then
subsequently call \code{\link[=fcds_load]{fcds_load()}} to load the cached, pre-processed data rather
than repeating the importing process or needing to locate the original
raw data. The data caching can be disabled by setting \code{output_file = NULL}.
}
\section{FCDS Data}{


Users must request the data from FCDS directly, via the
\href{https://fcds.med.miami.edu/inc/datarequest.shtml}{FCDS Data Request}
webpage.
}

\section{FCDS Recoding}{


This section will discuss the formatting for the FCDS recoding yaml file.
}

\references{
\url{https://fcds.med.miami.edu/}
}
\seealso{
Other FCDS Import Functions: \code{\link{fcds_cache}},
  \code{\link{fcds_default_data_path}},
  \code{\link{fcds_load}}
}
\concept{FCDS Import Functions}
