% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{fcds_map_add_label}
\alias{fcds_map_add_label}
\alias{fcds_map}
\alias{fcds_map_leaflet}
\title{Map FCDS Incidence Data}
\usage{
fcds_map_add_label(data, ...)

fcds_map(data, ..., fill = rate, geom_sf.color = "grey20",
  geom_sf.size = 0.25, coord_sf_hide = TRUE,
  theme = ggplot2::theme_minimal(),
  scale_fill = ggplot2::scale_fill_viridis_c())

fcds_map_leaflet(data, ..., palette = "Blues", palette_bins = if
  (length(palette == 1)) 5L else length(palette),
  group_name = "year_group",
  group_labels = rev(fcds_const(group_name)), proxy_id = NULL)
}
\arguments{
\item{data}{A data frame containing FCDS data. At a minimum either
\code{county_name} or \code{county_fips} are required to be included in the data. If
the input data is not an \code{sf} data frame, an attempt will be made to join
the data with the Florida county boundaries via \code{\link[=join_boundaries_fl]{join_boundaries_fl()}}.}

\item{...}{Ignored in \code{fcds_map()}.

Additional arguments passed to \code{\link[leaflet:leaflet]{leaflet::leaflet()}} in
\code{fcds_map_leaflet()}.

Additional arguments passed to \code{\link[glue:glue_data]{glue::glue_data()}} in
\code{fcds_map_add_label()}.}

\item{fill}{The unquoted column name to be used for the fill aesthetic.
Defaults to \code{rate}.}

\item{geom_sf.color}{The color of the \code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}} shape, typically
the color of the county outline.}

\item{geom_sf.size}{The size of the \code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}} shape, typically the
size of the county outline.}

\item{coord_sf_hide}{If \code{TRUE} (the default), removes lat/long guides.}

\item{theme}{The theme to be applied to the plot, by default
\code{\link[ggplot2:theme_minimal]{ggplot2::theme_minimal()}}. Set to \code{NULL} to use the global ggplot2 default
theme.}

\item{scale_fill}{The fill scale to be applied to the \code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}}
shape. By default \code{\link[ggplot2:scale_fill_viridis_c]{ggplot2::scale_fill_viridis_c()}}. Set equal to \code{NULL} to
skip the addition of this scale.}

\item{palette}{Set the color palette of the shape fill. You may provide a
vector of colors that will be used for the palette. The number of colors
provided determines the number of bins for the values mapped to the fill,
unless specified by \code{palette_bins}.}

\item{palette_bins}{The number of bins in the color \code{palette}.}

\item{group_name}{Character name of the column in \code{data} containing group
layers.}

\item{group_labels}{Labels for the group levels in the column referenced by
\code{group_name}. These labels should match the values in \code{data[[group_name]]}.}

\item{proxy_id}{If provided, a \code{\link[leaflet:leafletProxy]{leaflet::leafletProxy()}} map is created
instead. For use in Shiny apps.}
}
\value{
A ggplot2 plot object
}
\description{
Two interfaces are provided to mapping funcitonality in R. \code{fcds_map()} uses
\code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}} and \code{fcds_map_leaflet()} uses \code{\link[leaflet:leaflet]{leaflet::leaflet()}} to
create static and interactive plots of the input data. The defaults of both
functions are designed to work with \strong{fcds} data, in particular after an
\link[=age_adjust]{age-adjusted rate} has been calculated.
}
\section{Functions}{
\itemize{
\item \code{fcds_map_add_label}: Add a column with the name "\code{label}" to the input data.
The \code{...} are passed to \code{\link[glue:glue_data]{glue::glue_data()}}, so you can construct labels
row-wise with references to the column in \code{data}.

\item \code{fcds_map_leaflet}: Creates an interactive map using \code{\link[leaflet:leaflet]{leaflet::leaflet()}}.
}}

\examples{
fcds_example_rates <-
  fcds::fcds_example \%>\%
  count_fcds(moffitt_catchment = TRUE) \%>\%
  complete_age_groups() \%>\%
  age_adjust() \%>\%
  dplyr::ungroup() \%>\%
  dplyr::mutate(rate = rate / 5) \%>\%
  dplyr::mutate(year_group = factor(year_group, unique(year_group)))

fcds_map(fcds_example_rates) +
  ggplot2::facet_wrap(~year_group, ncol = 4) +
  ggplot2::theme(legend.position = c(0.93, 0.25))

\dontrun{
fcds_map_leaflet(fcds_example_rates)
}
}
