% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fcds_recoding}
\alias{fcds_recoding}
\title{List Available FCDS Recoding Settings Files}
\usage{
fcds_recoding(version = as.integer(strftime(Sys.Date(), "\%Y")))
}
\arguments{
\item{version}{The full name of the FCDS recoding settings file, or the year
of the release. If the year provided is more recent than the latest
tagged setting, the current release (package default) will be returned.}
}
\description{
Lists the available FCDS recoding settings files for use with
\code{\link[=fcds_import]{fcds_import()}}. Recoding settings are tagged with \code{-release-2018} indicating
the year of the FCDS release for which they are designed, except for the
latest release settings, which does not have this tag.
}
\examples{
fcds_recoding()
fcds_recoding(NULL)

fcds_recoding(2018)
fcds_recoding("fcds_recoding_release-2018")
fcds_recoding(2019)

}
\seealso{
Other FCDS Import Functions: \code{\link{fcds_cache}},
  \code{\link{fcds_default_data_path}},
  \code{\link{fcds_import}}, \code{\link{fcds_load}}
}
\concept{FCDS Import Functions}
