% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcds.R
\name{fcds_vars}
\alias{fcds_vars}
\title{Select Common FCDS Variable Groups}
\usage{
fcds_vars(..., .data = NULL)
}
\arguments{
\item{...}{Variable group names, as characters. Partial matching is allowed
and the matching is not case-sensitive. Possible variable groups include
\code{"id"}, \code{"demographics"}, \code{"cancer"}, \code{"icdo3"}, \code{"population"}, \code{"seer"},
\code{"tobacco"}.

A group may be input multiple times, but only the first appearance will be
used. The final variable list is ordered by the requested group, unless
\code{.data} is supplied.}

\item{.data}{If `.data`` is included, then the variables in the input data
are subset to only those appearing in the selected groups, in the order
they appear in the original data.}
}
\value{
A character vector of column names, or, if \code{.data} is provided, a
data frame subset to include columns matching the requested groups.
}
\description{
The FCDS data set includes a number of thematically-groups variables. This
function helps select those groups by returning variable names.
}
\examples{
fcds_vars("demo")
fcds_vars("demographics")

fcds_vars("id", "demo", "pop")

fcds_vars("seer", "tobacco", "seer")

tibble(
  patient_id = 1:5,
  year = 2000,
  county_name = "Pinellas",
  age_group = c("65 - 69", "10 - 14", "25 - 29", "70 - 74", "40 - 44"),
  cancer_status = c("Evidence of tumor", "Unknown", "No evidence of tumor",
                    "No evidence of tumor", "No evidence of tumor")
) \%>\%
  fcds_vars(.data = ., "id", "pop", "demo")

}
