% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{format_age_groups}
\alias{format_age_groups}
\title{Format Age Min/Max Boundaries into Standard Age Groups}
\usage{
format_age_groups(data, age_min = age_min, age_max = age_max,
  into = "age_group", missing_age_group = "Unknown",
  skip_overlap_check = FALSE)
}
\arguments{
\item{data}{A data frame containing the columns indicated by \code{age_min} and
\code{age_max}.}

\item{age_min}{The minimum age (inclusive) of the age group. May be \code{-Inf}
or \code{0}.}

\item{age_max}{The maximum age (inclusive) of the age group. May be \code{Inf}.}

\item{into}{The column name as a character string where the age group labels
will be stored in the input data frame.}

\item{missing_age_group}{The value for any age group with a missing boundary.}

\item{skip_overlap_check}{If \code{TRUE}, will skip the check ensuring that age
group upper boundaries do not match age group lower boundaries for any
age group other that those that are singular ages.}
}
\value{
A data frame with an additional column containing the age group
labels, named according to \code{into}.
}
\description{
This function formats \code{age_min} and \code{age_max} column into a new column, by
default called \code{age_group}, with a formatted string containing the age group
label denotes by the age boundaries.
}
\section{Age Formatting Guidelines}{
 Ages are formatted according to the
following rules.
\enumerate{
\item Ages range from \code{0} to \code{Inf}. Any age below \code{0} will be set to \code{0}.
Setting an age min boundary to \code{-Inf} is a resonable way of denoting \emph{all
ages under \code{age_max}}, but will result in age groups starting with \code{0}.
\item Age groups with an upper boundary of \code{Inf} will become "\code{{age_min}+}".
For example, if \code{age_min} is 85, then the age group label is \code{85+}. Use
\code{Inf} to denote all ages greater than or equal to \code{age_min}.
\item Single age groups are allowed, but there cannot otherwise be overlap
between the age boundaries.
\item Age boundaries are whole integers. Any partial ages are rounded down to
the highest integer less than the age boundary.
\item Missing age boundaries on either side result in an age group label of
\code{missing_age_group}, which is by default \code{"Unknown"}.
}
}

\examples{
d_age_group <- dplyr::tibble(
  age_min = c(-Inf, 20, 85),
  age_max = c(19, 84, Inf)
)

format_age_groups(d_age_group)

# format_age_groups() is the inverse of separate_age_groups()
d_age_group \%>\%
  format_age_groups() \%>\%
  dplyr::select(age_group) \%>\%
  separate_age_groups()

}
\seealso{
Other age processors: \code{\link{complete_age_groups}},
  \code{\link{filter_age_groups}},
  \code{\link{recode_age_groups}},
  \code{\link{separate_age_groups}},
  \code{\link{standardize_age_groups}}
}
\concept{age processors}
