% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcds-package.R
\docType{data}
\name{seer_fl_pop}
\alias{seer_fl_pop}
\title{SEER Florida population data}
\format{A data frame with 334273 rows and 11 variables: \describe{
\item{\code{year}}{Year} \item{\code{state}}{State postal abbreviation}
\item{\code{state_fips}}{State FIPS code} \item{\code{county_fips}}{County FIPS code}
\item{\code{registry}}{Registry} \item{\code{race}}{Race (White, Black, Other)}
\item{\code{origin}}{Origin: Non-Hispanic, Hispanic. Not applicable for data
prior to 1990.} \item{\code{sex}}{Sex} \item{\code{age_group}}{Age: 18 groups from 0,
0- 4, 5 - 9, \dots 85+} \item{\code{population}}{Population}
\item{\code{county_name}}{County Name, see \code{\link[=county_fips_code]{county_fips_code()}}} }}
\usage{
seer_fl_pop
}
\description{
Florida county-level population estimates from 1969--2016 in 18 age groups,
with races "white", "black", and "other". Downloaded from SEER on 2019-01-24
(see references for direct links). Recoding of column values was performed
according to the guidelines available at
\url{https://seer.cancer.gov/popdata/popdic.html}. For general consistency, all
variables are stored as \code{character} vectors. The original data includes 19
age groups, but age \code{0} is included with ages \code{1 - 4} in the \code{0 - 4} group
in this data.
}
\references{
\url{https://seer.cancer.gov/popdata/}

\url{https://seer.cancer.gov/popdata/yr1969_2016.19ages/fl.1969_2016.19ages.txt.gz}

\url{https://seer.cancer.gov/popdata/popdic.html}
}
\seealso{
Other SEER Population Data: \code{\link{seer_fl_pop_exp_race}}
}
\concept{SEER Population Data}
\keyword{datasets}
