% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcds-data.R
\docType{data}
\name{seer_icd_o_3}
\alias{seer_icd_o_3}
\title{SEER ICD-O-3 Histology Code and Behavior}
\format{A data frame with 12,084 rows and 4 columns
\describe{
\item{\code{histology}}{Histology Code}
\item{\code{histology_description}}{Description of the Histology Code (all upper case)}
\item{\code{histology_behavior}}{Histology Behavior Code, e.g. "8000/3"}
\item{\code{histology_behavior_description}}{Descrtipion of the Histology Behavior Code}
}}
\usage{
seer_icd_o_3
}
\description{
Contains the histology codes and behavior description provided by the
\href{https://seer.cancer.gov/icd-o-3/}{ICD-O-3 SEER Site/Histology Validation List}.
}
\references{
\url{https://seer.cancer.gov/icd-o-3/}
}
\keyword{datasets}
