% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcds-data.R
\docType{data}
\name{seer_std_ages}
\alias{seer_std_ages}
\title{SEER Standard Population for Age-Adjustment}
\format{A data frame with 18 rows and 4 variables:
\describe{
\item{\code{standard}}{Standard}
\item{\code{standard_name}}{Standard Name}
\item{\code{age_group}}{Age group}
\item{\code{std_pop}}{Standard Population}
}}
\usage{
seer_std_ages
}
\description{
US. Standard populations, year 2000, for age-adjusted statistics. Downloaded
from SEER on 2019-01-24 (see References for direct links). This data set is
the \emph{2000 U.S. Std Population (single ages to 84 - Census P25-1130)}, with
ages agregated into 18 age groups: "\code{0 - 4}", ..., "\code{80 - 84}", \code{\link{85+}}.
}
\references{
\url{https://seer.cancer.gov/stdpopulations/}

\url{https://seer.cancer.gov/stdpopulations/stdpop.singleagesthru84.txt}
}
\keyword{datasets}
