% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{separate_age_groups}
\alias{separate_age_groups}
\title{Separate Age Group Column into Low and High Age of Range}
\usage{
separate_age_groups(data, age_group = age_group, sep = "\\\\s*-\\\\s*",
  ..., age_low = age_low, age_high = age_high)
}
\arguments{
\item{data}{A data frame.}

\item{age_group}{Unquoted column name containing the age grouping.}

\item{sep}{Separator between columns.

If character, is interpreted as a regular expression. The default
value is a regular expression that matches any sequence of
non-alphanumeric values.

If numeric, interpreted as positions to split at. Positive values start
at 1 at the far-left of the string; negative value start at -1 at the
far-right of the string. The length of \code{sep} should be one less than
\code{into}.}

\item{...}{Not used other than to require explicit naming of arguments.}
}
\description{
Takes the age group stored in \code{age_group} and creates two variables,
\code{age_low} and \code{age_high} with the age boundaries of the group. By default,
\code{separate_age_groups()} assumes that the age group definitions are separated
by a dash (\code{"-"}), possibly with whitespace on either side. The separator
can be specified using the \code{sep} argument.
}
\examples{
d_age_group <- dplyr::tibble(
  id = 1:4,
  age_group = c("0 - 4", "10 - 14", "65 - 69", "85+")
)

separate_age_groups(d_age_group)

}
\seealso{
Other age processors: \code{\link{complete_age_groups}},
  \code{\link{filter_age_groups}},
  \code{\link{standardize_age_groups}}
}
\concept{age processors}
