% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year.R
\name{separate_year_groups}
\alias{separate_year_groups}
\title{Separate Year Groups}
\usage{
separate_year_groups(data, year = year, sep = "-")
}
\arguments{
\item{data}{A data frame}

\item{year}{The column containing the \code{year_group}.}

\item{sep}{Characters separating years in \code{year}. Whitespace on either side
of \code{sep} will be automatically removed. Passed to \code{\link[tidyr:separate]{tidyr::separate()}}.}
}
\description{
Separates a column containing year groups, such as "\code{1983-1985}", into
two separate columns with "\code{_min}" and "\code{_max}" appended to the column names.
}
\seealso{
Other year processors: \code{\link{add_mid_year_groups}},
  \code{\link{complete_year_groups}}
}
\concept{year processors}
