% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{standardize_age_groups}
\alias{standardize_age_groups}
\title{Standardize Age Groups}
\usage{
standardize_age_groups(data = NULL, age_group = age_group,
  std_age_groups = fcds_const("age_group"), ...)
}
\arguments{
\item{data}{A data frame.}

\item{age_group}{The column containing age or age group. This column will be
overwritten if named \code{age_group}. If the column is numeric or can be
coerced to numeric, it is treated as actual age. Otherwise it is assumed to
contain age groups.}

\item{std_age_groups}{Standard age groups, in the desired order.}

\item{...}{Arguments passed on to \code{separate_age_groups}
\describe{
  \item{data}{A data frame.}
  \item{age_group}{Unquoted column name containing the age grouping.}
  \item{sep}{Separator between columns.

If character, is interpreted as a regular expression. The default
value is a regular expression that matches any sequence of
non-alphanumeric values.

If numeric, interpreted as positions to split at. Positive values start
at 1 at the far-left of the string; negative value start at -1 at the
far-right of the string. The length of \code{sep} should be one less than
\code{into}.}
}}
}
\description{
Standardizes age groups to match \link{seer_std_ages}. Adds an \code{age_group} column,
or overwites the existing \code{age_group} column if present. The \code{age_var} column
can be either an actual age or an age group. If the column is numeric, it
assumed to be an actual age, otherwise if it contains any non-numeric
characters it is assumed to be an age group that will be expanded using
\code{\link[=separate_age_groups]{separate_age_groups()}}. If multiple standardized age groups match a given age,
the function will throw an error.
}
\examples{
dplyr::tibble(
  id = 1:4,
  age_group = c("0 - 4", "10-14", "65-69", "85+")
) \%>\%
  standardize_age_groups()
}
\seealso{
Other age processors: \code{\link{complete_age_groups}},
  \code{\link{filter_age_groups}},
  \code{\link{separate_age_groups}}
}
\concept{age processors}
