% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year.R
\name{add_mid_year_groups}
\alias{add_mid_year_groups}
\title{Add Mid-Year of Year Group}
\usage{
add_mid_year_groups(data, year = year, sep = "-", offset = NULL)
}
\arguments{
\item{data}{A data frame}

\item{year}{The column containing the \code{year_group}.}

\item{sep}{Characters separating years in \code{year}. Whitespace on either side
of \code{sep} will be automatically removed. Passed to \code{\link[tidyr:separate]{tidyr::separate()}}.}

\item{offset}{If supplied, the number of years to be added to the lower bound
of the year group to calculate the mid-year value. By default uses
\code{floor((year_max - year_min) / 2)}.}
}
\description{
Adds a new column containing the midpoint year of the range given in the
column indicated by the \code{year} argument, in a new column with "\code{_mid}"
appended to the \code{year} column name. The \code{year} column is assumed to contain
a \emph{min} year and a \emph{max} year, separated by \code{sep}. The midpoint is calculated
as \code{floor((year_max - year_min) / 2)} unless \code{offset} is explicitly provided.
}
\examples{
dplyr::tibble(year = fcds_const("year")) \%>\%
  add_mid_year_groups()

}
\seealso{
Other year processors: \code{\link{complete_year_groups}},
  \code{\link{separate_year_groups}}
}
\concept{year processors}
