% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age_adjust}
\alias{age_adjust}
\title{Age-Adjust Rates}
\usage{
age_adjust(data, count = n, population = fcds::seer_pop_fl,
  population_standard = fcds::seer_std_ages, by_year = c(year_mid =
  "year"), age = age_group, keep_age = FALSE)
}
\arguments{
\item{data}{A data frame, containing counts}

\item{count}{The unquoted column name containing raw event or outcome counts.}

\item{population}{Population data specific to the population described by
\code{data}. By default, uses the county-specific Florida population data
provided by SEER (see \link{seer_pop_fl}).}

\item{population_standard}{The standard age-specific population used to
calculate the age-adjusted rates. By default, uses the 2000 U.S. standard
population provided by SEER (see \link{seer_std_ages}).}

\item{keep_age}{Age-adjustment by definition summarizes event or outcome
counts (incidence) over \emph{all} included ages. Set \code{keep_age = TRUE} to join
the source data with all age-specific population without completing the age
adjustment calculation. This option is primarily provided for debugging
purposes.}

\item{year}{The unquoted column name containing the year that will be matched
to the population data.}
}
\value{
A data frame with age-adjusted incidence rates in the column \code{rate}.
Note that the \code{age} column will no longer be included in the output because
the age-adjusted rate summarizes the observed incidence across all ages.

If \code{keep_age} is \code{TRUE}, the age column is retained, but the final rate is
not calculated, adding the columns \code{population}, \code{std_pop}, and \code{w} for the
specific population, standard population and standardizing population
weight, respectively.
}
\description{
Provides age-adjusted incident rates with respect to a reference population,
by default \link{seer_std_ages}.
}
\section{Age-Adjusted Rates}{


Calculating age-adjusted rates requires three primary inputs:
\enumerate{
\item Raw age-specific count of event or outcome, possibly observed or
summarized at repeated, consistent time intervals.
\item Population data with the same demographic resolution as the age-specific
counts, as in, for example, the population for the same geographic region,
sex, race, year, and age.
\item The standard reference population that is used to weight incidence among
the observed age-specific count.
}

Each input is required to store age information. The default data supplied
with the package for \code{population} (\link{seer_pop_fl}) and \code{population_standard}
(\link{seer_std_ages}) use the column name \code{age_group}. You can specify the name
of the column containing age information with the \code{age} argument. If the
column name in \code{data} is not present in the population data, \code{age_adjust()}
will fall back to \code{age_group} for those data sets.
}

\examples{

# This example is drawn from https://seer.cancer.gov/seerstat/tutorials/aarates/
d_incidence <- tibble::tribble(
  ~age_group,   ~n,
     "0 - 4",  116,
     "5 - 9",   67,
   "10 - 14",   71,
   "15 - 19",   87,
   "20 - 24",  177,
   "25 - 29",  290,
   "30 - 34",  657,
   "35 - 39", 1072,
   "40 - 44", 1691,
   "45 - 49", 2428,
   "50 - 54", 2931,
   "55 - 59", 2881,
   "60 - 64", 2817,
   "65 - 69", 2817,
   "70 - 74", 2744,
   "75 - 79", 2634,
   "80 - 84", 1884,
       "85+", 1705
) \%>\%
  dplyr::mutate(year_mid = 2013) \%>\%
  standardize_age_groups()

d_population <- tibble::tribble(
  ~age_group, ~population,
     "0 - 4",      693068,
     "5 - 9",      736212,
   "10 - 14",      770999,
   "15 - 19",      651390,
   "20 - 24",      639159,
   "25 - 29",      676354,
   "30 - 34",      736557,
   "35 - 39",      724826,
   "40 - 44",      700200,
   "45 - 49",      617437,
   "50 - 54",      516541,
   "55 - 59",      361170,
   "60 - 64",      259440,
   "65 - 69",      206204,
   "70 - 74",      172087,
   "75 - 79",      142958,
   "80 - 84",       99654,
       "85+",       92692,
) \%>\%
  dplyr::mutate(year = 2013) \%>\%
  standardize_age_groups()

# Because the example data do not include the year of observation, we set
# by_year = NULL so that age_adjust() does not attempt to join
# d_incidence with d_population by a year column.

age_adjust(d_incidence, population = d_population, by_year = NULL)

age_adjust(d_incidence,
           population = d_population,
           by_year = NULL,
           keep_age = TRUE)

}
\references{
\url{https://seer.cancer.gov/seerstat/tutorials/aarates/}
}
