% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{complete_age_groups}
\alias{complete_age_groups}
\title{Complete Age Groups}
\usage{
complete_age_groups(data, age_low = NULL, age_high = NULL, ...,
  age_var = age_group, fill = list(n = 0), include_unknown = FALSE,
  std_age_groups = fcds_const("age_group"))
}
\arguments{
\item{data}{A data frame.}

\item{age_low}{Youngest age (inclusive).}

\item{age_high}{Eldest age (inclusive).}

\item{...}{If \code{age_low} and \code{age_high} are missing from}

\item{age_var}{Unquoted column name containing the age grouping.}

\item{fill}{A named list that for each variable supplies a single value to
use instead of \code{NA} for missing combinations.}

\item{include_unknown}{Should the "Unknown" age group be included?}
}
\description{
Completes age groups by adding missing age groups, either within the age
range from \code{age_low} to \code{age_high} or using the full age list from
\link{seer_std_ages}.
}
\examples{
tibble::tibble(
  age_group = c("10 - 14", "15 - 19", "25 - 29"),
  n = 10:12
) \%>\%
  complete_age_groups(10, 35)

}
\seealso{
Other age processors: \code{\link{expand_age_groups}},
  \code{\link{filter_age}},
  \code{\link{standardize_age_groups}}
}
\concept{age processors}
