% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcds.R
\name{count_fcds}
\alias{count_fcds}
\title{Count FCDS Cases}
\usage{
count_fcds(data, ..., sex = NULL, race = NULL, origin = NULL,
  county_name = NULL, default_groups = c("year_group", "year",
  "age_group"), discard_unseen_levels = TRUE,
  moffitt_catchment = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{...}{Unquoted column names to be added to the grouping of the output
and subsequent counting.}

\item{sex}{Character vector of values of \code{sex} to be included in count, or
\code{TRUE} to include all values of \code{sex} present in input data}

\item{race}{Character vector of values of \code{race} to be included in count, or
\code{TRUE} to include all values of \code{race} present in input data}

\item{origin}{Character vector of values of \code{origin} to be included in count,
or \code{TRUE} to include all values of \code{origin} present in input data}

\item{county_name}{Character vector of values of \code{county_name} to be included
in count, or \code{TRUE} to include all values of \code{county_name} present in the
input data, or \code{"moffitt_catchment"} to limit to the counties in the
catchment area of the \href{https://moffitt.org}{Moffitt Cancer Center}.}

\item{default_groups}{Variables that should be included in the grouping,
prior to counting cancer cases. Set to \code{NULL} to use only the groups
already present in the input data.}

\item{discard_unseen_levels}{If \code{TRUE}, drops factor levels that don't appear
in the results. If \code{FALSE}, then no changes are made to the factor levels.
If a character string of column names, then only unobserved levels in those
columns are dropped.}

\item{moffitt_catchment}{\strong{Deprecated.} Please use \code{county_name = "moffitt_catchment"} instead to limit counties to those in the catchment
area of the \href{https://moffitt.org}{Moffitt Cancer Center}.}
}
\value{
A grouped data frame with counts. The output groups includes the
union of the groups of the original input \code{data}, the groups specified by
the columns indicated in \code{...}, and the \code{default_groups} added by
\code{count_fcds()} (modifyable by the \code{default_groups} argument).

All factor levels will be modified to include only those levels that appear
in the final output across all groups.
}
\description{
Helper function to count cancer cases in the FCDS data by year and age group,
in addition to any groups already present in the data. For convenience, you
may additionally filter to include particular values of sex, race, year,
county name and hispanic ethnicity. See \code{\link[=fcds_const]{fcds_const()}} for more information
about possible values for these variables. By default, \code{count_fcds()} ensures
that \code{age_group}, \code{year_group}, and \code{year} are included in the grouping
variables if they are present in the data. If they are not, or if they are
not present in the FCDS data, then it would be better to use \code{\link[dplyr:count]{dplyr::count()}}
directly.
}
\examples{

fcds_example \%>\%
  dplyr::filter(county_name == "Pinellas") \%>\%
  count_fcds(cancer_site_group, sex = "Male", county_name = TRUE)

fcds_example \%>\%
  filter_age_groups(age_gt = 20, age_lt = 25) \%>\%
  count_fcds(sex = TRUE, county_name = c("Pinellas", "Hillsborough"))

}
