% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcds-package.R
\docType{package}
\name{fcds-package}
\alias{fcds}
\alias{fcds-package}
\title{fcds: Process data from the Florida Cancer Data System}
\description{
Helper functions for processing data from the
    Florida Cancer Data System.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GerkeLab/fcds}
  \item Report bugs at \url{https://github.com/GerkeLab/fcds/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{Garrick.Aden-Buie@moffitt.org}

Authors:
\itemize{
  \item Jordan Creed \email{Jordan.H.Creed@moffitt.org}
  \item Travis Gerke \email{Travis.Gerke@moffitt.org}
}

}
\keyword{internal}
