% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcds.R
\name{filter_age}
\alias{filter_age}
\title{Filter Data by Age Range}
\usage{
filter_age(.data, age_low = 0, age_high = Inf, age_var = age_group)
}
\arguments{
\item{.data}{A data frame}

\item{age_low}{Youngest age (inclusive)}

\item{age_high}{Eldest age (inclusive)}

\item{age_var}{Unquoted column name containing the age grouping.}
}
\description{
Filters data to include persons with ages in the range between \code{age_low} and
\code{age_high}.
}
\seealso{
Other age processors: \code{\link{complete_age_groups}},
  \code{\link{expand_age_groups}}
}
\concept{age processors}
