% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{filter_age_groups}
\alias{filter_age_groups}
\title{Filter Data by Age Range}
\usage{
filter_age_groups(data, age_low = 0, age_high = Inf,
  age_group = age_group)
}
\arguments{
\item{data}{A data frame.}

\item{age_low}{Youngest age (inclusive).}

\item{age_high}{Eldest age (inclusive).}

\item{age_group}{Unquoted column name containing the age grouping.}
}
\description{
Filters data to include persons with ages in the range between \code{age_low} and
\code{age_high}. If \code{age_group} has not been expanded into low and high ages of
the range, the input data is first passed to \code{\link[=expand_age_groups]{expand_age_groups()}}. If the
boundary age lies within a group, that group is \emph{not included} in the output.
}
\examples{
d_age_group <- tibble::tibble(
  id = 1:4,
  age_group = c("0 - 4", "10 - 14", "65 - 69", "85+")
)

d_age_group \%>\%
  filter_age_groups(age_low = 0, age_high = 15)

d_age_group \%>\%
  filter_age_groups(age_low = 65)

# Notice that the "65 - 69" group is *not* included
d_age_group \%>\%
  filter_age_groups(age_high = 66)

}
\seealso{
Other age processors: \code{\link{complete_age_groups}},
  \code{\link{expand_age_groups}},
  \code{\link{standardize_age_groups}}
}
\concept{age processors}
