% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{join_boundaries_fl}
\alias{join_boundaries_fl}
\title{Join Florida County Boundaries to a data frame}
\usage{
join_boundaries_fl(data, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Additional arguments passed to \code{\link[USAboundaries:us_counties]{USAboundaries::us_counties()}}.}
}
\value{
An \code{sf} data frame, created by \code{\link[sp:st_sf]{sp::st_sf()}}.
}
\description{
Uses the \code{sf} and \code{USAboundaries} packages to join the input data with the
county shapes required for spatial data visualization. See
\code{\link[USAboundaries:us_counties]{USAboundaries::us_counties()}} for more information about the county
shapes used.
}
