% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcds-data.R
\docType{data}
\name{seer_icd_o_3}
\alias{seer_icd_o_3}
\title{SEER ICD-O-3 Histology Code and Behavior}
\format{A data frame with 12,084 rows and 5 variables:
\describe{
\item{\code{site_group}}{integer. Site group, e.g. "\code{Esophagus}"}
\item{\code{histology}}{character. Four-digit ICD-O-3 histology code, e.g. "\code{8262}"}
\item{\code{histology_description}}{integer. Description of histology, e.g. "\code{Papillary Adenocarcinoma, NOS}"}
\item{\code{morphology}}{character. Five-digit ICD-O-3 morphology code, e.g. "\code{8262/3}"}
\item{\code{morphology_description}}{integer. Description of morphology, e.g. "\code{Villous Adenocarcinoma}"}
}}
\usage{
seer_icd_o_3
}
\description{
Contains the histology codes and behavior description provided by the
\href{https://seer.cancer.gov/icd-o-3/}{ICD-O-3 SEER Site/Histology Validation List}.
}
\references{
\url{https://seer.cancer.gov/icd-o-3/}
}
\keyword{datasets}
