% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year.R
\name{separate_year_groups}
\alias{separate_year_groups}
\title{Separate Year Groups}
\usage{
separate_year_groups(data, year_group = year_group,
  into = c("year_min", "year_max"), sep = "-")
}
\arguments{
\item{data}{A data frame}

\item{year_group}{The unquoted column containing the \code{year_group}.}

\item{into}{The names of the min year and max year columns (respectively)
into which the minimum and maximum year from \code{year_group} are added. Set
to \code{NULL} to use the \code{year_group} column name as a prefix with \code{_min} and
\code{_max} appended.}

\item{sep}{Characters separating years in \code{year}. Whitespace on either side
of \code{sep} will be automatically removed. Passed to \code{\link[tidyr:separate]{tidyr::separate()}}.}
}
\description{
Separates a column containing year groups, such as "\code{1983-1985}", into
two separate columns named \code{year_min} and \code{year_max} by default.
}
\examples{
year_groups <- c("1981-1985", "1986-1990", "1991-1995", "1996-2000",
                 "2001-2005", "2006-2010", "2011-2015")

dplyr::tibble(year_group = year_groups) \%>\%
  separate_year_groups()

}
\seealso{
Other year processors: \code{\link{add_mid_year_groups}},
  \code{\link{complete_year_groups}}
}
\concept{year processors}
